refactor - use separate file for batches endpoints
